Fix binary asynchronous request
