+++ lib/XMLHttpRequest.js
@@ -109,6 +109,7 @@ exports.XMLHttpRequest = function () {
 
   // Result & response
   this.responseText = ''
+  this.responseBinary = []
   this.responseXML = ''
   this.status = null
   this.statusText = null
@@ -432,29 +433,58 @@ exports.XMLHttpRequest = function () {
           return
         }
 
-        response.setEncoding('utf8')
+        const encoding = responseType === 'text' ? 'utf8' : 'binary'
+        if (encoding === 'utf8') {
+          response.setEncoding('utf8')
+        }
 
         setState(self.HEADERS_RECEIVED)
         self.status = response.statusCode
 
-        response.on('data', function (chunk) {
-          // Make sure there's some data
-          if (chunk) {
-            self.responseText += chunk
-          }
-          // Don't emit state changes if the connection has been aborted.
-          if (sendFlag) {
-            setState(self.LOADING)
-          }
-        })
+        if (encoding === 'utf8') {
+          response.on('data', function (chunk) {
+            // console.log('unxhr data utf8', chunk.length)
+            // Make sure there's some data
+            if (chunk) {
+              self.responseText += chunk
+            }
+            // Don't emit state changes if the connection has been aborted.
+            if (sendFlag) {
+              setState(self.LOADING)
+            }
+          })
 
-        response.on('end', function () {
-          if (sendFlag) {
-            // Discard the end event if the connection has been aborted
-            setState(self.DONE)
-            sendFlag = false
-          }
-        })
+          response.on('end', function () {
+            // console.log('unxhr sendFlag', sendFlag)
+            if (sendFlag) {
+              // Discard the end event if the connection has been aborted
+              setState(self.DONE)
+              sendFlag = false
+            }
+          })
+        } else {
+          response.on('data', function (chunk) {
+            // console.log('unxhr data binary', chunk.length)
+            // Make sure there's some data
+            if (chunk) {
+              self.responseBinary.push(chunk)
+            }
+            // Don't emit state changes if the connection has been aborted.
+            if (sendFlag) {
+              setState(self.LOADING)
+            }
+          })
+
+          response.on('end', function () {
+            // buffers are Uint8Array instances
+            self.response = Buffer.concat(self.responseBinary).buffer
+            if (sendFlag) {
+              // Discard the end event if the connection has been aborted
+              setState(self.DONE)
+              sendFlag = false
+            }
+          })
+        }
 
         response.on('error', function (error) {
           self.handleError(error)
